Got rid of the package structure in favor of a flat structure
The package structure was causing path issues
